add/delete employee wasn't functional

+++ src/main/java/igorts2004/staffer/domain/Employee.java
@@ -13,23 +13,23 @@ public class Employee {
 	@Id
     @Column(name = "id")
     @GeneratedValue	
-	private int id;
+	private long id;
 	
 	@Column(name = "name")
 	private String name;
 	
 	public Employee() {}
 	
-	public Employee(int id, String name) {
+	public Employee(long id, String name) {
 		this.id = id;
 		this.name = name;
 	}
 
-	public int getId() {
+	public long getId() {
 		return id;
 	}
 
-	public void setId(int id) {
+	public void setId(long id) {
 		this.id = id;
 	}
